Session#5 - Level Up Your AI Game: Build AI Agents with Google ADK/Trip Planner And Recipe Agents/requirements.txt
google-adk
requests
python-dotenv
litellm